+++ R/chooseClusterNumber.R
@@ -3,9 +3,11 @@
 #' For a subset of the data, perform clustering under a range of cluster numbers.
 #'  Report on loglikelihood vs. number of clusters, and suggest a best choice.
 #' @param counts Counts matrix, cells * genes. 
-#' @param neg Vector of mean negprobe counts per cell
+#' @param neg Vector of mean negprobe counts per cell (default = "rna")
+#' @param assay_type Assay type of RNA, protein 
 #' @param bg Expected background
 #' @param fixed_profiles Matrix of cluster profiles to hold unchanged throughout iterations.
+#' @param fixed_sds Matrix of SDs expression of genes x cell types,to hold unchanged throughout iterations. Only for assay_type of protein
 #' @param cohort Vector of cells' cohort assignments. 
 #' @param init_clust Vector of initial cluster assignments.
 #' @param n_clusts Vector giving a range of cluster numbers to consider.
@@ -41,6 +43,7 @@
 chooseClusterNumber <-
   function(counts,
            neg,
+           assay_type = c("rna", "protein"),
            bg = NULL,
            fixed_profiles = NULL,
            fixed_sds = NULL,
@@ -52,10 +55,10 @@ chooseClusterNumber <-
            align_genes = TRUE,
            plotresults = FALSE,
            nb_size = 10,
-           assay_type,
            pct_drop = 0.005,
            min_prob_increase = 0.05,
            ...) {
+    assay_type <- match.arg(tolower(assay_type), c("rna", "protein"))  
 
   # infer bg if not provided: assume background is proportional to the scaling factor s
   s <- rowSums(counts)